requirements.txt
glob2==0.7
torch==1.12.0
torchvision==0.13.0
opencv-python==4.1.2.30
numpy==1.22.3
Pillow==9.0.1
scipy==1.8.0
scikit-image==0.19.2
scikit-learn==0.24.2
lpips==0.1.4
